backend/requirements.txt
Flask==3.1.2
Flask_Cors==4.0.0
python-dotenv==1.2.1
Requests==2.32.5
boto3